# Use `Vec` in `SingleEpochParticipationCache` rather than `HashMap`