## Issue Addressed

Which issue # does this PR address?

## Proposed Changes

Uses a `Vec` in `SingleEpochParticipationCache` rather than `HashMap` to speed up processing times at the cost of memory usage.

### Benchmarks

Benchmarks on a recent mainnet state using #3252 to get timing.

#### Without this PR

```
lcli skip-slots --state-path /tmp/state-0x3cdc.ssz --partial-state-advance --slots 32 --state-root 0x3cdc33cd02713d8d6cc33a6dbe2d3a5bf9af1d357de0d175a403496486ff845e --runs 10
[2022-06-09T08:21:03Z INFO  lcli::skip_slots] Run 0: 245.718794ms
[2022-06-09T08:21:03Z INFO  lcli::skip_slots] Run 1: 245.364782ms
[2022-06-09T08:21:03Z INFO  lcli::skip_slots] Run 2: 255.866179ms
[2022-06-09T08:21:04Z INFO  lcli::skip_slots] Run 3: 243.838909ms
[2022-06-09T08:21:04Z INFO  lcli::skip_slots] Run 4: 250.431425ms
[2022-06-09T08:21:04Z INFO  lcli::skip_slots] Run 5: 248.68765ms
[2022-06-09T08:21:04Z INFO  lcli::skip_slots] Run 6: 262.051113ms
[2022-06-09T08:21:05Z INFO  lcli::skip_slots] Run 7: 264.293967ms
[2022-06-09T08:21:05Z INFO  lcli::skip_slots] Run 8: 293.202007ms
[2022-06-09T08:21:05Z INFO  lcli::skip_slots] Run 9: 264.552017ms
```

#### With this PR:

```
lcli skip-slots --state-path /tmp/state-0x3cdc.ssz --partial-state-advance --slots 32 --state-root 0x3cdc33cd02713d8d6cc33a6dbe2d3a5bf9af1d357de0d175a403496486ff845e --runs 10
[2022-06-09T08:24:05Z INFO  lcli::skip_slots] Run 0: 126.377026ms
[2022-06-09T08:24:05Z INFO  lcli::skip_slots] Run 1: 134.491018ms
[2022-06-09T08:24:05Z INFO  lcli::skip_slots] Run 2: 127.580876ms
[2022-06-09T08:24:05Z INFO  lcli::skip_slots] Run 3: 134.919047ms
[2022-06-09T08:24:06Z INFO  lcli::skip_slots] Run 4: 128.497402ms
[2022-06-09T08:24:06Z INFO  lcli::skip_slots] Run 5: 126.495564ms
[2022-06-09T08:24:06Z INFO  lcli::skip_slots] Run 6: 126.949893ms
[2022-06-09T08:24:06Z INFO  lcli::skip_slots] Run 7: 127.155165ms
[2022-06-09T08:24:06Z INFO  lcli::skip_slots] Run 8: 130.017356ms
[2022-06-09T08:24:06Z INFO  lcli::skip_slots] Run 9: 130.155124ms
```

## TODO

- [ ] Quantify additional memory costs.
- [ ] Fix `todo(paul)` in code